move context fixture to conftest for reuse
