+++ tests/test_context.py
@@ -1,20 +1,6 @@
 import json
 
-import pytest
-from mirror_builder import context
 from packaging.requirements import Requirement
-
-
-@pytest.fixture
-def tmp_context(tmp_path):
-    ctx = context.WorkContext(
-        tmp_path / 'sdists-repo',
-        tmp_path / 'wheels-repo',
-        tmp_path / 'work-dir',
-        0,
-    )
-    ctx.setup()
-    return ctx
 
 
 def test_seen(tmp_context):
